Move SingleFileHTMLBuilder to sphinx.builders.singlehtml package
